terrain_gen/requirements.txt
torch
transformers
opencv-python
numpy
trimesh
Pillow
